finalproject-main/requirements.txt
hume
python-dotenv
groq
requests
Pillow
PyQt5
PyQtWebEngine
PyQtChart
hume[microphone]
python-dotenv

